updated: actions 'use server' fix

+++ web-app/src/app/blog/actions/deletePost.ts
@@ -1,7 +1,14 @@
+"use server";
+
 // DB
 import { createClient } from "@/utils/supabase/server";
 
-export async function deletePost(id: string) {
+export async function deletePost(formData: FormData) {
+  const id = formData.get("id")?.toString();
+  if (!id) {
+    console.warn("Post ID is missing.");
+    return;
+  }
   const supabase = await createClient();
 
   const { error } = await supabase.from("posts").delete().eq("id", id);
